+++ tests/integration/go/basic/main.go
@@ -1,0 +1,26 @@
+package main
+
+import (
+	corev1 "github.com/pulumi/pulumi-kubernetes/sdk/go/kubernetes/core/v1"
+	"github.com/pulumi/pulumi/sdk/go/pulumi"
+)
+
+func main() {
+	pulumi.Run(func(ctx *pulumi.Context) error {
+		_, err := corev1.NewPod(ctx, "foo", &corev1.PodArgs{
+			Spec: corev1.PodSpecArgs{
+				Containers: corev1.ContainerArray{
+					corev1.ContainerArgs{
+						Name:  pulumi.String("nginx"),
+						Image: pulumi.String("nginx"),
+					},
+				},
+			},
+		})
+		if err != nil {
+			return err
+		}
+
+		return nil
+	})
+}