+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 ## HEAD (Unreleased)
 
+### Improvements
+
+-   Add a Go SDK. (https://github.com/pulumi/pulumi-kubernetes/pull/1029).
+
 ### Bug fixes
 
 -   Update the Python `Provider` class to use parameter naming consistent with other resources. (https://github.com/pulumi/pulumi-kubernetes/pull/1039).
+-   Change URN for apiregistration resources. (https://github.com/pulumi/pulumi-kubernetes/pull/1021).
 
 ## 1.5.8 (March 16, 2020)